Pull some interfaces into higher levels to avoid iterating on the fly
